Formatting stage first letter
